backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
PyMySQL
python-jose
passlib[bcrypt]
redis
